pyproject.toml
[project]
name = "lpil-pygmentizer"
version = "0.0.1"
description = "A lpil specific wrapper around the Pygments library."
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
    "pygments>=2.16.1",
]
requires-python = ">=3.10"
license = {text = "Apache2.0"}

[project.scripts]
lpilPygmentize = "lpilPygments.cli:cli"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

